Core, Core.UI, Core.UI.WPF - organize UI libraries
