+++ SoluiNet.DevTools.Core/api-ui-wpf/.gitignore
@@ -1,0 +1,5 @@
+###############
+#  temp file  #
+###############
+*.yml
+.manifest
