+++ SoluiNet.DevTools.Core/api-ui-wpf/index.md
@@ -1,0 +1,2 @@
+# PLACEHOLDER
+TODO: Add .NET projects to the *src* folder and run `docfx` to generate **REAL** *API Documentation*!
