+++ SoluiNet.DevTools.Core/api-ui/.gitignore
@@ -1,0 +1,5 @@
+###############
+#  temp file  #
+###############
+*.yml
+.manifest
