+++ SoluiNet.DevTools.Core/images/.gitignore
@@ -1,0 +1,2 @@
+/logo.svg
+/favicon.ico
